Feedback, make tests unit tests
